Updated to correct comments
Corrected comments of Shengtao and Fangli - factors of 2pi are now included in all drives. Omega has been changed to be a nice round number in units of 2pi, and time has been scaled slightly in the nonequilibrium dynamics example. NOTE THAT these dynamics are very slightly wrong, due to an incorrect C6 strength, with a difference of 4/pi, but that should not change very much.