rework hamiltonian (#175)
this PR reworks the entire representation and split the old representation into 2 layers IR, the first layer is YaoBlocks, the second is `Hamiltonian` which stores the list of time-dependent prefactor and list of constant terms. Then `StepHamiltonian` is treated as a linear-map that supports `LinearAlgebra.mul!` interface. So that we can use it inside krylov and ODE methods.